// packages/bench/bebop/generated.ts
export interface Vec3 {
	readonly x: number;

	readonly y: number;

	readonly z: number;
}

export const Vec3 = /*#__PURE__*/ Object.freeze(
	/*#__PURE__*/ Object.assign(
		// Factory function
		(data: Vec3): Vec3 & BebopRecord => {
			return Object.freeze({
				...data,
				encode(): Uint8Array {
					return Vec3.encode(this);
				},
			});
		},
		// Static methods
		{
			encode(record: Vec3): Uint8Array {
				const view = BebopView.getInstance();
				view.startWriting();
				Vec3.encodeInto(record, view);
				return view.toArray();
			},

			encodeInto(record: Vec3, view: BebopView): void {
				view.writeFloat32(record.x);
				view.writeFloat32(record.y);
				view.writeFloat32(record.z);
			},

			decode(buffer: Uint8Array): Vec3 & BebopRecord {
				const view = BebopView.getInstance();
				view.startReading(buffer);
				const decoded = Vec3.readFrom(view);
				return Vec3(decoded);
			},

			readFrom(view: BebopView): Vec3 {
				let field0: number;
				field0 = view.readFloat32();
				let field1: number;
				field1 = view.readFloat32();
				let field2: number;
				field2 = view.readFloat32();
				return {
					x: field0,
					y: field1,
					z: field2,
				};
			},
		},
	),
);

export interface Player {
	readonly id: number;

	readonly pos: Vec3;

	readonly vel: Vec3;

	readonly health: number;

	readonly score: number;

	readonly team: number;
}

export const Player = /*#__PURE__*/ Object.freeze(
	/*#__PURE__*/ Object.assign(
		// Factory function
		(data: Player): Player & BebopRecord => {
			return Object.freeze({
				...data,
				encode(): Uint8Array {
					return Player.encode(this);
				},
			});
		},
		// Static methods
		{
			encode(record: Player): Uint8Array {
				const view = BebopView.getInstance();
				view.startWriting();
				Player.encodeInto(record, view);
				return view.toArray();
			},

			encodeInto(record: Player, view: BebopView): void {
				view.writeUint32(record.id);
				Vec3.encodeInto(record.pos, view);
				Vec3.encodeInto(record.vel, view);
				view.writeUint16(record.health);
				view.writeUint32(record.score);
				view.writeByte(record.team);
			},

			decode(buffer: Uint8Array): Player & BebopRecord {
				const view = BebopView.getInstance();
				view.startReading(buffer);
				const decoded = Player.readFrom(view);
				return Player(decoded);
			},

			readFrom(view: BebopView): Player {
				let field0: number;
				field0 = view.readUint32();
				let field1: Vec3;
				field1 = Vec3.readFrom(view);
				let field2: Vec3;
				field2 = Vec3.readFrom(view);
				let field3: number;
				field3 = view.readUint16();
				let field4: number;
				field4 = view.readUint32();
				let field5: number;
				field5 = view.readByte();
				return {
					id: field0,
					pos: field1,
					vel: field2,
					health: field3,
					score: field4,
					team: field5,
				};
			},
		},
	),
);

export interface GameTick {
	readonly tick: number;

	readonly dt: number;

	readonly flags: number;

	readonly players: Player[];
}

export const GameTick = /*#__PURE__*/ Object.freeze(
	/*#__PURE__*/ Object.assign(
		// Factory function
		(data: GameTick): GameTick & BebopRecord => {
			return Object.freeze({
				...data,
				encode(): Uint8Array {
					return GameTick.encode(this);
				},
			});
		},
		// Static methods
		{
			encode(record: GameTick): Uint8Array {
				const view = BebopView.getInstance();
				view.startWriting();
				GameTick.encodeInto(record, view);
				return view.toArray();
			},

			encodeInto(record: GameTick, view: BebopView): void {
				view.writeUint32(record.tick);
				view.writeFloat32(record.dt);
				view.writeByte(record.flags);
				{
					const length0 = record.players.length;
					view.writeUint32(length0);
					for (let i0 = 0; i0 < length0; i0++) {
						Player.encodeInto(record.players[i0], view);
					}
				}
			},

			decode(buffer: Uint8Array): GameTick & BebopRecord {
				const view = BebopView.getInstance();
				view.startReading(buffer);
				const decoded = GameTick.readFrom(view);
				return GameTick(decoded);
			},

			readFrom(view: BebopView): GameTick {
				let field0: number;
				field0 = view.readUint32();
				let field1: number;
				field1 = view.readFloat32();
				let field2: number;
				field2 = view.readByte();
				let field3: Player[];
				{
					const length0 = view.readUint32();
					field3 = [];
					for (let i0 = 0; i0 < length0; i0++) {
						let x0: Player;
						x0 = Player.readFrom(view);
						field3[i0] = x0;
					}
				}
				return {
					tick: field0,
					dt: field1,
					flags: field2,
					players: field3,
				};
			},
		},
	),
);